Extract prerequisites for running the integration and acceptance tests
