+++ pkg/acceptance/testing.go
@@ -200,19 +200,19 @@ func TestAccPreCheck(t *testing.T) {
 		_, _ = SecondaryTestClient().Schema.CreateTestSchemaIfNotExists(t)
 		_, _ = SecondaryTestClient().Warehouse.CreateTestWarehouseIfNotExists(t)
 
-		if err := helpers.EnsureQuotedIdentifiersIgnoreCaseIsSetToFalse(atc.client, ctx); err != nil {
+		if err := TestClient().EnsureQuotedIdentifiersIgnoreCaseIsSetToFalse(ctx); err != nil {
 			t.Fatal(err)
 		}
 
-		if err := helpers.EnsureQuotedIdentifiersIgnoreCaseIsSetToFalse(atc.secondaryClient, ctx); err != nil {
+		if err := SecondaryTestClient().EnsureQuotedIdentifiersIgnoreCaseIsSetToFalse(ctx); err != nil {
 			t.Fatal(err)
 		}
 
-		if err := helpers.EnsureScimProvisionerRolesExist(atc.client, ctx); err != nil {
+		if err := TestClient().EnsureScimProvisionerRolesExist(ctx); err != nil {
 			t.Fatal(err)
 		}
 
-		if err := helpers.EnsureScimProvisionerRolesExist(atc.secondaryClient, ctx); err != nil {
+		if err := SecondaryTestClient().EnsureScimProvisionerRolesExist(ctx); err != nil {
 			t.Fatal(err)
 		}
 	})
@@ -232,11 +232,6 @@ func ConfigurationDirectory(directory string) func(config.TestStepConfigRequest)
 	return func(req config.TestStepConfigRequest) string {
 		return filepath.Join("testdata", directory)
 	}
-}
-
-func DefaultConfig(t *testing.T) *gosnowflake.Config {
-	t.Helper()
-	return atc.config
 }
 
 func TestClient() *helpers.TestClient {
